+++ workers/workers.go
@@ -142,6 +142,7 @@ func (w *Workers) Stop() {
 }
 
 type Job struct {
+	count     int
 	tasks     chan func() error
 	completed chan struct{}
 	result    chan error
@@ -171,6 +172,10 @@ func (j *Job) Wait() error {
 	return <-j.result
 }
 
+func (j *Job) Workers() int {
+	return j.count
+}
+
 // NewJob creates a new job and adds it to the workers' queue. [taskBacklog]
 // specifies the maximum number of tasks that can be added to the created job
 // channel.
@@ -185,6 +190,7 @@ func (w *Workers) NewJob(taskBacklog int) (*Job, error) {
 		return nil, ErrShutdown
 	}
 	j := &Job{
+		count:     w.count,
 		tasks:     make(chan func() error, taskBacklog),
 		completed: make(chan struct{}),
 		result:    make(chan error, 1),
